shiro-demo/shiro-boot-demo/src/test/java/com/github/viqbgrg/shirobootdemo/controller/UserLoginControllerTest.java
package com.github.viqbgrg.shirobootdemo.controller;

import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;

import static org.assertj.core.api.Assertions.*;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@WebMvcTest(UserLoginController.class)
class UserLoginControllerTest {
    @Autowired
    private MockMvc mvc;
    @Test
    void signIn() throws Exception {
        this.mvc.perform(post("/login/vehicle").accept(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(content().string("Honda Civic"));
    }
}